Filters class restructured into multiple classes; added VT100 filter
