+++ expectit-core/src/main/java/net/sf/expectit/filter/Filters.java
@@ -30,35 +30,9 @@ import java.util.regex.Pattern;
  */
 public final class Filters {
 
-    private static final String COLORS_REGEXP_STRING = "\\x1b\\[([0-9]{1,2}(;[0-9]{1,2})?)?[m|K]";
-    /**
-     * The regular expression which matches <a href="http://en.wikipedia
-     * .org/wiki/ANSI_escape_code#Colors">ANSI escape
-     * sequences for colors</a>.
-     */
-    public static final Pattern COLORS_PATTERN = Pattern.compile(COLORS_REGEXP_STRING);
-
-    /**
-     * The regular expression which matches non printable characters: {@code
-     * [\x00\x08\x0B\x0C\x0E-\x1F]}.
-     */
-    public static final Pattern NON_PRINTABLE_PATTERN = Pattern.compile(
-            "[\\x00\\x08\\x0B\\x0C\\x0E-\\x1F]");
-
     private static final int DEFAULT_FILTER_OVERLAP = 80;
 
     private Filters() {
-    }
-
-    /**
-     * Creates a filter which removes all the non-printable characters matching {@link
-     * #NON_PRINTABLE_PATTERN} in the
-     * input string.
-     *
-     * @return the filter
-     */
-    public static Filter removeNonPrintable() {
-        return replaceInBuffer(NON_PRINTABLE_PATTERN, "");
     }
 
     /**
@@ -92,17 +66,6 @@ public final class Filters {
      */
     public static Filter replaceInString(final String regexp, final String replacement) {
         return replaceInString(Pattern.compile(regexp), replacement);
-    }
-
-    /**
-     * Creates a filter which removes <a href="http://en.wikipedia
-     * .org/wiki/ANSI_escape_code#Colors">ANSI
-     * escape sequences for colors</a> in the input.
-     *
-     * @return the filter
-     */
-    public static Filter removeColors() {
-        return replaceInBuffer(COLORS_PATTERN, "");
     }
 
     /**
@@ -176,52 +139,5 @@ public final class Filters {
      */
     public static Filter replaceInBuffer(final Pattern regexp, final String replacement) {
         return replaceInBuffer(regexp, replacement, DEFAULT_FILTER_OVERLAP);
-    }
-
-    /**
-     * Combines the filters in a filter chain.
-     * <p/>
-     * The given filters are applied one by one in the order that hey appear in the method
-     * argument list.
-     * <p/>
-     * The string returns by the
-     * {@link Filter#beforeAppend(String, StringBuilder)} method of one filter is passed a
-     * parameter to the next one if it is not {@code null}. If it is {@code null},
-     * then the {@code beforeAppend}
-     * won't be called any more and the latest non-null result is appended to the expect internal
-     * buffer.
-     * <p/>
-     * If the return value of the {@link Filter#afterAppend(StringBuilder)} method is true,
-     * then all the calls
-     * of this method on the consequent filters will be suppressed.
-     *
-     * @param filters the filters, not {@code null}
-     * @return the combined filter
-     */
-    public static Filter chain(final Filter... filters) {
-        return new FilterAdapter() {
-            @Override
-            protected String doBeforeAppend(String string, StringBuilder buffer) {
-                String previousResult = null;
-                for (Filter filter : filters) {
-                    string = filter.beforeAppend(string, buffer);
-                    if (string == null) {
-                        return previousResult;
-                    }
-                    previousResult = string;
-                }
-                return string;
-            }
-
-            @Override
-            protected boolean doAfterAppend(StringBuilder buffer) {
-                for (Filter filter : filters) {
-                    if (filter.afterAppend(buffer)) {
-                        return true;
-                    }
-                }
-                return false;
-            }
-        };
     }
 }
