+++ expectit-core/src/test/java/net/sf/expectit/KarafExample.java
@@ -20,9 +20,9 @@ package net.sf.expectit;
  * #L%
  */
 
-import static net.sf.expectit.filter.Filters.chain;
-import static net.sf.expectit.filter.Filters.removeColors;
-import static net.sf.expectit.filter.Filters.removeNonPrintable;
+import static net.sf.expectit.filter.FilterChain.chain;
+import static net.sf.expectit.filter.BuiltinFilters.removeColors;
+import static net.sf.expectit.filter.BuiltinFilters.removeNonPrintable;
 import static net.sf.expectit.matcher.Matchers.contains;
 import static net.sf.expectit.matcher.Matchers.eof;
 import static net.sf.expectit.matcher.Matchers.regexp;
